Create Temperature Converter
This is a program in which we take input Temperature value from the user and then ask the unit of the temperature entered. After that we ask from the user in which unit of Temperature he/she wants to convert. Then accordingly we give the desired output in desired unit.