Update Calculator Program
This is a program in which we take input from the user about the operation they want to perform.  The user is given 5 options to choose from which is add, multiply, subtract, divide and exit. when the user will enter 'exit', at once the program will end, otherwise, it will keep asking for the two numbers with which they want to perform the above operations.