To swap elements in an array
In this program, we first take input from the user about the length of the array it needs. Then, we ask to input numbers and thus we create the array. After creating an array we ask for the indexes of the two elements of the array that the user wants to swap. Then, by using functions or methods we swap those elements. Then, we display the newly created array.