To Reverse an Array
In this program, we create an array according to user input like the length of the array and its elements of it. After using methods in java, we reverse the array by letting it undergo a certain procedure under "for loop". Then, we print the reversed array as an output.