Linear Search in Array
In this program, we create an array according to the length and elements inputted by the user. Using methods in java, we run a for loop to perform the linear search, if the number that the user wants to search for is found then it returns "Yes" with the Index of it.