To Find Maximum or Minimum Value in an Array
In this program, we create an array according to length and elements inputted by the user. After that user needs to give a command if he/she wants to find the Minimum or Maximum value in that array. If either of the commands is given, we by using methods in java with for-loop, if statement and letting the array undergo a procedure and then give output as the Maximum or Minimum value.