To do Linear search for a number in 2D array
In this program, we create a 2D array as per the data inputted by the user. After that by using for-loop in methods in java, and by letting the 2D array undergo certain procedures in the loop, we check if the number is there or not. If the number is found then it prints the indexes like in which row the number is and in that row on which column.