To find minimum or maximum value in an array
In this program, we create a 2D array according to the data inputted by the user. After that, we ask for input from the user if it wants to find a Minimum value or Maximum value. We create separate methods for each operation like for finding the minimum or maximum value. After getting the request as either of the options, we then run the method designed for that. Thus, at last, we print the minimum or maximum value in the array as the output of the program.