To find no. of digits in elements of array & no. of elements with even no. of digits
In this program, we create an array according to data inputted by the user. After that, we print the array Of Number of digits in each element of the array. Then, the Number of elements having an even number of digits in the array. At last, we print the array of elements that have an even number of digits.