Richest Customer Wealth Leetcode Question
In this program, we have given the solution to a leetcode question. In that, we need to first add individual elements of an individual array in a 2D array.  Then, we have to find the maximum value among the sums of elements of an individual arrays in a 2D array.