Order Agnostic Binary Search In Array
In this program, we first create an Array according to the data inputted by the user. After that, we ask for the number user wants to check for. Then we use methods in java, in methods we check if the array given is in ascending order or descending. After that, we use the algorithm accordingly in the method body.  Thus, we give the index of the number as output if it exists else we return -1.