To find ceiling of a number in Array
In this program, we create an array according to the data inputted by the user. We then ask for a number of which user wants to find the ceiling in the array. Note. The ceiling of a number is a number that is the smallest greater number in the array.  So, by using binary search we give the output.