To sort an array using Selection Sort Algrorithm
In this program, we first create an array according to the user input. Then we use the Selection sort algorithm to sort the given array. Which we do by implementing it in a method and then calling it with the inputted array as input to the method that sorts. Finally, it outputs the sorted array.