To sort an array using Cycle Sort Algorithm
In this program, we first create an array according to the user input. Then we use the  Cycle sort algorithm to sort the given array. Which we do by implementing it in a method and then calling it with the inputted array as input to the method that sorts. Finally, it outputs the sorted array. Please note that only 1 to N numbers are allowed in the array that is going to be sorted using the cycle sort algorithm. Also, the common difference between the values should be one only. Otherwise, it will give an error.