Doing Binary Search Using Recursion
In this program we first create an array using inputs from the user like the length of the array and the elements, then we ask for the target element from the user. Using this data, we apply the binary search in a function or method in java, where we use recursion, by the formula, O(1) + SearchingSpace(N/2) each time. At last printing the output as an index of the target element.