To Find if the number is Even or Odd using Bitwise Operator
In this program, we first get the number from the user as input. After that, using the property of AND(&) operator, which is if you apply & operator on a number with 1, then it is going to return the Least Significant Bit, so if that bit is 1 then the number should be odd else it should be even. So, using this in a Method, that returns true if odd and false if even. Then, we print "The number is Odd" if true, else - "The number is Even".