To Find Nth Magic Number
In this program, we first take input from the user as "N", to output the "Nth" Magic Number. A Magic Number is a sum of the multiplication of each bit with 5 raised to the power of the position of that bit. So, we make a function or Method to do so. Thus, giving output as a Magic Number at Nth place.