To Count Bits Of A Number
In this program, we take any number as input. Then by using a Method to count its bits. That is to use a counter inside the loop and keep increasing it by 1, with it doing the right shift of bits by 1 at the same time, till the number becomes zero. You can see that in the "Count" Method or Function. Thus, giving the number of bits in a Number as the output.