To Count Number Of Set Bits In Any Number
In this Program, we take any number as an input. Then using Bitwise operators we determine the number of set bits in that number. I mean in a while loop until the number is greater than 0. Doing "And" Operation with "1" to get Least significant bit and checking if it is setbit that is if it is "1". Then, if it is Set Bit then we add 1 to the counter. Then We use the "Right Shift" operator and do n right shift by 1, which will delete the least significant bit. Thus, at last, getting the Output as a number of Set Bits in that number.