To find XOR in a Range From A to B
In this program, we first take input from the user as A and B as a Range. Then, First, we find XOR from "0" To "B". Then, in order to delete XOR till "A-1", we do XOR between XOR from "0" to "B" and XOR from "0" to "A-1". Thus, at last, getting the XOR From Range A to B as Output.