To print prime numbers upto N using Sieve Of Eratosthenes
In this program, we take any Number up to which the user wants to know primes. Then, we first create an array of booleans having all values as false. Using the Sieve Of Eratosthenes approach, we create a method and turn all the non-prime indexes of that array to true. Then we print the indexes of the array with "false" values. Thus giving the output as all the primes Upto N.