Selection Sort Using Recursion
In this program, we first create an array using the user's inputs. Then in Method or Function, we perform our algorithm using Recursion. We first assume the element at the 0th index as maximum then we traverse through the array to check if it is the maximum element. If that is the max then we swap it with the last element of the array.