To Print Subsets of an Array
In this program, we first create an array using inputs from the user. Then we apply an algorithm that is we create an outer list of lists and an inner list, for each num in the array we add that num in the for each list in outer lists. To avoid duplication we check for each num if it is the same as the previous num, if that's so we start our loop after that num, that is previous end + 1;
In the end, we get a list of lists containing all the subsets of that array.