BackTracking Problem - N-Queens
In this program we take input as a Number, then A row number. We first make a Number by Number boolean matrix.  Then we make a Recursive method or Function. Which places a queen in one place and then checks if N queens can be placed in the next [row][col]s and if it can be placed we mark our boolean matrix at that index as true. if we were able to place the queen at the last row of that matrix, we display our boolean matrix. If not then we backtrack and while going to the above function calls we undo our work and mark true as false. And continue our Recursive tree likewise.