Introduction to Objects And Classes
In this program, we create an object of class Students n times and print data inside it. "n" is inputted by the user and inside the for loop that runs n times, it asks for the name, roll number, and marks of that particular student that it stores as an object which is an instance of class "Students" and then prints its data one by one.