To create Tick-Tack-Toe Game
We first create an empty 3 by 3 matrix, then we define two players "x" and "o", we first start with x. We use the while loop until the variable "Gameover" becomes true. Then we, take a row and column input from the player, and after that, we check if that place is empty, if not empty we reject the input. If that place is empty then we proceed by placing "x" there. Now, we check if that player has won by using a function, if yes then the "Gameover" variable will be "true" and the loop will exit. if have not won then we will switch the players, like here to "o".